refactoring and bugfixes for bbusters / mechatt
