+++ src/mame/arcade.flt
@@ -751,6 +751,7 @@ mcr3.cpp
 mcr68.cpp
 meadows.cpp
 meadwttl.cpp
+mechatt.cpp
 mediagx.cpp
 megadriv_acbl.cpp
 megaphx.cpp
